Setup RabbitMQ (#4)
* Setup RabbitMQ

* Clean up gRPC server and add test coverage

* Support build updates and improve pid management

* Update Bottle SHA

* Expand on build attributes

Add model and order id to all builds. Additionally, support the INPROGRESS
status.

* Ignore /lib

* Update Cachex usage

* Add logging

* Emit message on build updates

* Add handler logging

* Improvements

* Use hal_id for logs

* Ignore messages from self

* Update builds to work off HAL ids

* Add missing component support

Since this is easy to calculate we won't persist it by rather keep it in
state. Following this we can update the gRPC call to return the missing
components for display in the UI.

* Update gRPC payload and data source

Add support for missing components, update our inquiries into builds
from the DB to the supervisor

* Logging updates

* Ensure missing components is populated

* Remove failing test

This test wasn't great to begin with and now it's resulting in a suite
failure. We should devise a better test for this.